Fix RSC boundary violation: move productItemTransform out of client module
page.tsx (Server Component) was calling productItemTransform directly,
but it was exported from product-viewed/index.tsx, a 'use client' file.
Next.js treats every export of a 'use client' module as part of the
client boundary, so calling it from the server threw at runtime:
"Attempted to call productItemTransform() from the server but
productItemTransform is on the client."

Moved the transform to its own plain module (no 'use client' needed,
it's just a pure data mapper) so both the client component and the
server page can import it safely.
