+++ core/app/[locale]/(default)/product/[slug]/_components/product-viewed/index.tsx
@@ -6,25 +6,11 @@ import { FragmentOf } from '~/client/graphql';
 import { bodl } from '~/lib/bodl';
 
 import { ProductViewedFragment } from './fragment';
+import { productItemTransform } from './transform';
 
 interface Props {
   product: FragmentOf<typeof ProductViewedFragment>;
 }
-
-export const productItemTransform = (p: FragmentOf<typeof ProductViewedFragment>) => {
-  return {
-    product_id: p.entityId.toString(),
-    product_name: p.name,
-    brand_name: p.brand?.name,
-    sku: p.sku,
-    sale_price: p.prices?.salePrice?.value,
-    purchase_price: p.prices?.salePrice?.value || p.prices?.price.value || 0,
-    base_price: p.prices?.price.value,
-    retail_price: p.prices?.retailPrice?.value,
-    currency: p.prices?.price.currencyCode || 'USD',
-    variant_id: p.variants.edges?.map((variant) => variant.node.entityId),
-  };
-};
 
 export const ProductViewed = ({ product }: Props) => {
   useEffect(() => {
