+++ core/app/[locale]/(default)/product/[slug]/page.tsx
@@ -16,7 +16,8 @@ import { client } from '~/client';
 
 import { addToCart } from './_actions/add-to-cart';
 import { ProductSchema } from './_components/product-schema';
-import { ProductViewed, productItemTransform } from './_components/product-viewed';
+import { ProductViewed } from './_components/product-viewed';
+import { productItemTransform } from './_components/product-viewed/transform';
 import { PaginationSearchParamNames, Reviews } from './_components/reviews';
 import { QnAList } from './_components/qna-list';
 import { getProductData } from './page-data';
